backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.24,<1.0
httpx>=0.27,<1.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
mangum>=0.17.0
